Binkp implemetation based on Apache Mina.

+++ binkp/src/main/java/org/fidonet/binkp/BinkpConnector.java
@@ -145,7 +145,7 @@ public class BinkpConnector implements ProtocolConnector {
     }
 
     private BinkpCommand getCommand(int command) {
-        return BinkpCommand.findCommand(command);
+        return BinkpCommand.findCommand((byte)command);
     }
 
     /**
